Rename file EditingWorkflow > EditingWorkflowImpl
