+++ src/packages/editing/EditingWorkflowImpl.test.tsx
@@ -12,7 +12,7 @@ import { createServiceOptions, setupMap, waitForMapMount } from "@open-pioneer/m
 import { PackageContextProvider } from "@open-pioneer/test-utils/react";
 import { render } from "@testing-library/react";
 import { PackageIntl, Service } from "@open-pioneer/runtime";
-import { EditingWorkflowImpl } from "./EditingWorkflow";
+import { EditingWorkflowImpl } from "./EditingWorkflowImpl";
 import BaseLayer from "ol/layer/Base";
 import { Interaction } from "ol/interaction";
 import { MapRegistryImpl } from "@open-pioneer/map/MapRegistryImpl";
